+++ src/main/java/lib/controllers/ProfileXboxGamepad.kt
@@ -1,5 +1,7 @@
 package lib.controllers
 
+import lib.profiles.DriverProfile
+import lib.profiles.defaultDriverProfile
 import kotlin.math.pow
 
 /**
